datastreams: move OpenAIREProjectHTTPReader to generic OpenAIREHTTPReader
